requirements.txt
build
hatch
pytest
pytest-watcher
openai
python-dotenv
requests
